refactor: reestructuración de DTOs, mappers y lógica de controladores
- Creación de nuevos DTOs específicos para operaciones de recursos (ej: UpdateProfileDTO, CreateProductDTO) para mejorar la tipado y validación.
- Actualización de mappers para soportar los nuevos DTOs y facilitar la conversión de entidades.
- Migración de la lógica de transformación de DTOs desde los servicios hacia los controladores, dejando los servicios enfocados exclusivamente en la manipulación de entidades.
- Limpieza y estandarización del código heredado para alinear el estilo y la convención.

+++ src/main/java/com/techlab/store/dto/ReviewDTO.java
@@ -8,7 +8,7 @@ public record ReviewDTO(
         Long id,
         String reviewerName,
         String reviewerEmail,
-        Integer rating,
+        Double rating,
         String comment,
         Long productId
 ) { }
